fix: import problems for linter and ts
