+++ .env
@@ -2,5 +2,4 @@ NODE_ENV=development
 PORT=4000
 AUTH_MODE=false
 JWT_SECRET_KEY=secret-key
-# Set your database connection information here
-MONGO_CONNECTION_STRING=your-mongo-db-connection-string
+DB_CONNECTION=my-db-connection-string
