+++ .eslintrc.json
@@ -14,23 +14,26 @@
     }
   },
   "settings": {
-    "noInlineConfig": true
+    "noInlineConfig": true,
+    "import/resolver": {
+      "typescript": {},
+      "node": {
+        "extensions": [".js", ".jsx", ".ts", ".tsx"]
+      }
+    }
   },
   "extends": [
-    "eslint:recommended",
-    "plugin:node/recommended",
-    "plugin:@typescript-eslint/recommended",
     "airbnb-base",
+    "eslint:recommended",
+    "plugin:import/typescript",
+    "plugin:@typescript-eslint/recommended",
     "prettier"
   ],
   "rules": {
-    "import/extensions": "off",
     "linebreak-style": "off",
-    "node/no-unsupported-features/es-syntax": "off",
+    "import/extensions": "off",
     "no-underscore-dangle": "off",
-    "import/prefer-default-export": "off"
-    // "@typescript-eslint/indent": ["error", 2],
-    // "@typescript-eslint/no-unused-vars": "error",
-    // "@typescript-eslint/no-explicit-any": "error"
+    "import/prefer-default-export": "off",
+    "node/no-unsupported-features/es-syntax": "off"
   }
 }
