+++ package.json
@@ -1,13 +1,13 @@
 {
   "name": "rs-rest-service-starter",
-  "version": "1.0.0",
+  "version": "1.0.1",
   "repository": {
     "type": "git",
     "url": ""
   },
   "type": "module",
   "description": "REST service starter for RS School",
-  "main": "src/server.js",
+  "main": "build/server.js",
   "scripts": {
     "build": "tsc --noEmitOnError",
     "start": "nodemon build/server.js",
@@ -54,18 +54,19 @@
     "yamljs": "0.3.0"
   },
   "devDependencies": {
+    "@types/express": "^4.17.12",
     "@types/node": "^15.6.1",
+    "@types/swagger-ui-express": "^4.1.2",
+    "@types/yamljs": "^0.2.31",
     "@typescript-eslint/eslint-plugin": "^4.25.0",
     "@typescript-eslint/parser": "^4.25.0",
-    "@types/swagger-ui-express": "^4.1.2",
-    "@types/express": "^4.17.12",
-    "@types/yamljs": "^0.2.31",
     "chai": "4.3.4",
     "debug": "4.3.1",
     "dirty-chai": "2.0.1",
     "eslint": "^7.2.0",
     "eslint-config-airbnb-base": "14.2.1",
     "eslint-config-prettier": "8.3.0",
+    "eslint-import-resolver-typescript": "^2.4.0",
     "eslint-plugin-import": "2.22.1",
     "eslint-plugin-node": "11.1.0",
     "eslint-plugin-prettier": "3.4.0",
