--- src/_server.ts
@@ -1,6 +1,0 @@
-// import { PORT } from './common/config.js';
-// import { app } from './app.js';
-
-// app.listen(PORT, () => {
-//   console.log(`Application is running on http://localhost:${PORT}`)
-// });
